start developing map page
